split aws.go into multiple files
